Lot of shuffling just for updateUI
