+++ swarm.cabal
@@ -666,6 +666,8 @@ library swarm-tui
     Swarm.TUI.Controller
     Swarm.TUI.Controller.MainEventHandler
     Swarm.TUI.Controller.REPLEventHandler
+    Swarm.TUI.Controller.SaveScenario
+    Swarm.TUI.Controller.UpdateUI
     Swarm.TUI.Controller.Util
     Swarm.TUI.Editor.Controller
     Swarm.TUI.Editor.Json
@@ -681,6 +683,7 @@ library swarm-tui
     Swarm.TUI.Launch.View
     Swarm.TUI.List
     Swarm.TUI.Model
+    Swarm.TUI.Model.Achievements
     Swarm.TUI.Model.Event
     Swarm.TUI.Model.Goal
     Swarm.TUI.Model.Menu
